agent/requirements.txt
pipecat-ai[daily,openai,deepgram,cartesia,twilio,silero]
fastapi
uvicorn
python-dotenv
aiohttp
twilio
python-multipart
